Add keyboard functionality to calculator
Calculator eventListeners for keyup events were not working as
calculator was a div which cannot receive focus by default. Could have put the eventListener on document but prefer to keep it specific in case future elements outside the calculator div was added. Now  clicking on the calculator element followed by button press works.